pyproject.toml
[tool.poetry]
name = "migra"
version = "3.0"
authors = [ "Robert Lechte <robertlechte@gmail.com>",]
license = "Unlicense"
readme = "README.md"
description = "Like `diff` but for PostgreSQL schemas"

repository = "https://github.com/djrobstep/migra"
homepage = "https://databaseci.com/docs/migra"

[tool.poetry.dependencies]
python = ">=3.7,<4"
sqlbag = "*"
six = "*"
# schemainspect = {path="../schemainspect", develop=true}
# schemainspect = ">=3.1.1663480743"
psycopg2-binary = { version="*", optional = true }

[tool.poetry.dev-dependencies]
sqlbag = "*"
pytest = "*"
pytest-cov = "*"
pytest-clarity = "*"
psycopg2-binary = "*"
flake8 = "*"
isort = "*"
black = "*"
toml = "*"
testcontainers-postgres = "*"
sqlalchemy = "1.4.50"

[tool.poetry.scripts]
migra = 'migra:do_command'

[tool.poetry.extras]
pg = ["psycopg2-binary"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
